We should branch after the first beta.
We did this for 1.6 and it was very effective. 95%+ of fixes which merge
during the alpha are backported, as the policy is "all but really major
features". It's easier to just not merge any really major features.
After beta, we have feature freeze so we need to backport bugs to stable
but not features, so then the branch makes sense.